# apps/dashboard/lib/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Carbon API",
    "description": "Control plane for Carbon — projects, ingestion, emulators, snapshots, artifacts, org/billing, and CLI auth. Every endpoint accepts either an API key (`x-carbon-key`) or a Better Auth session cookie/bearer token; the docs and `/health` endpoints are the only routes served anonymously.",
    "version": "dev",
    "contact": {
      "name": "Carbon",
      "url": "https://github.com/TaxCollector23/carbon",
      "email": "support@example.com"
    },
    "license": { "name": "Apache-2.0", "url": "https://www.apache.org/licenses/LICENSE-2.0" }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-carbon-key",
        "description": "A key minted via `POST /v1/api-keys` or the `bootstrap` script. Prefix `ck_live_*` for production, `ck_test_*` for test envs."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "better-auth-session",
        "description": "A Better Auth session bearer token. Preferred for programmatic access from a signed-in user."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "The Better Auth session cookie set by the dashboard. Used implicitly when the browser calls the API cross-origin with `credentials: \"include\"`."
      }
    },
...
  },
  "paths": {
...
    "/v1/events": {
      "get": {
        "summary": "List audit events",
        "tags": ["Events"],
        "description": "Return audit events for the caller's org in descending time order. Supports keyset pagination via `cursor` (ISO 8601 timestamp of the last item seen) and optional filtering by `projectId` or `action`.",
        "parameters": [
          {
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": { "type": "string", "format": "date-time" },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "query",
            "name": "projectId",
            "required": false
          },
          {
            "schema": { "type": "string", "minLength": 1, "maxLength": 120 },
            "in": "query",
            "name": "action",
            "required": false
          },
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "query",
            "name": "orgId",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "orgId": { "type": "string" },
                          "projectId": { "type": "string", "nullable": true },
                          "actorType": { "type": "string" },
                          "actorId": { "type": "string", "nullable": true },
                          "action": { "type": "string" },
                          "metadata": {},
                          "createdAt": { "type": "string", "format": "date-time" }
                        },
                        "required": [
                          "id",
                          "orgId",
                          "projectId",
                          "actorType",
                          "actorId",
                          "action",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": { "type": "string", "nullable": true },
                    "hasMore": { "type": "boolean" }
                  },
                  "required": ["data", "nextCursor", "hasMore"],
                  "additionalProperties": false
                },
                "example": {
                  "data": [
                    {
                      "id": "evt_01HXK5N9Q1B7C4D3E2F1G0H9J8",
                      "orgId": "org_01HXK5H7Q9C0R3Q1S8V6M4WJZK",
                      "projectId": "prj_01HXK5H7Q9C0R3Q1S8V6M4WJZK",
                      "actorType": "user",
                      "actorId": "usr_01HXK5H7Q9C0R3Q1S8V6M4WJZK",
                      "action": "project.created",
                      "metadata": { "slug": "checkout-api", "name": "Checkout API" },
                      "createdAt": "2025-11-14T18:22:41.000Z"
                    }
                  ],
                  "nextCursor": null,
                  "hasMore": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/events/stream": {
      "get": {
        "summary": "Server-Sent Events stream of audit events",
        "tags": ["Events"],
        "description": "Long-lived text/event-stream connection. Emits `hello` on open, `ping` heartbeats every ~15s, and `new-event` frames as events land for the caller's org.",
        "parameters": [
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "query",
            "name": "orgId",
            "required": false
          },
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "query",
            "name": "projectId",
            "required": false
          },
          {
            "schema": { "type": "string", "minLength": 1, "maxLength": 120 },
            "in": "query",
            "name": "action",
            "required": false
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/events/export": {
      "get": {
        "summary": "Export audit events as CSV",
        "tags": ["Export"],
        "description": "Stream audit events as a CSV attachment. Same filters as `GET /v1/events` but returns a single denormalized CSV row per event, capped at `limit` (max 10,000).",
        "parameters": [
          {
            "schema": { "type": "string", "enum": ["csv"], "default": "csv" },
            "in": "query",
            "name": "format",
            "required": false
          },
          {
            "schema": { "type": "integer", "minimum": 1, "maximum": 10000, "default": 1000 },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "query",
            "name": "projectId",
            "required": false
          },
          {
            "schema": { "type": "string", "minLength": 1, "maxLength": 120 },
            "in": "query",
            "name": "action",
            "required": false
          },
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "query",
            "name": "orgId",
            "required": false
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
  },
  "servers": [{ "url": "http://localhost:4000", "description": "local dev" }],
  "security": [{ "apiKey": [] }, { "bearerAuth": [] }, { "sessionCookie": [] }],
...
}